# admin/contracts/admin-payment.openapi.json
{
  "components": {
    "schemas": {
...
      "AdminUserSummaryOut": {
        "properties": {
          "address": {
            "title": "Address",
            "type": "string"
          },
          "asset": {
            "title": "Asset",
            "type": "string"
          },
          "chain_error": {
            "default": "",
            "title": "Chain Error",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "database_balance": {
            "title": "Database Balance",
            "type": "string"
          },
          "disabled": {
            "title": "Disabled",
            "type": "boolean"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "network": {
            "title": "Network",
            "type": "string"
          },
          "on_chain_balance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "On Chain Balance"
          },
          "telegram_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telegram User Id"
          },
          "telegram_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telegram Username"
          },
          "total_accounts_purchased": {
            "title": "Total Accounts Purchased",
            "type": "integer"
          },
          "total_amount_paid": {
            "title": "Total Amount Paid",
            "type": "string"
          },
          "trx_balance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trx Balance"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "display_name",
          "disabled",
          "address",
          "network",
          "asset",
          "database_balance",
          "currency",
          "total_accounts_purchased",
          "total_amount_paid"
        ],
        "title": "AdminUserSummaryOut",
        "type": "object"
      },
      "AdminUsersOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminUserSummaryOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "title": "AdminUsersOut",
        "type": "object"
      },
...
      "BalanceOut": {
        "properties": {
          "available_balance": {
            "title": "Available Balance",
            "type": "string"
          },
          "balance": {
            "title": "Balance",
            "type": "string"
          },
          "credit_limit": {
            "default": "0",
            "title": "Credit Limit",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          }
        },
        "required": [
          "balance",
          "available_balance",
          "currency"
        ],
        "title": "BalanceOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IssueUserRequest": {
        "properties": {
          "deposit_address": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 10,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deposit Address"
          },
          "display_name": {
            "default": "",
            "maxLength": 200,
            "title": "Display Name",
            "type": "string"
          },
          "telegram_user_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telegram User Id"
          },
          "telegram_username": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telegram Username"
          }
        },
        "title": "IssueUserRequest",
        "type": "object"
      },
      "IssuedUserOut": {
        "properties": {
          "api_key": {
            "title": "Api Key",
            "type": "string"
          },
          "asset": {
            "title": "Asset",
            "type": "string"
          },
          "deposit_address": {
            "title": "Deposit Address",
            "type": "string"
          },
          "network": {
            "title": "Network",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "api_key",
          "network",
          "asset",
          "deposit_address"
        ],
        "title": "IssuedUserOut",
        "type": "object"
      },
      "MeOut": {
        "properties": {
          "asset": {
            "title": "Asset",
            "type": "string"
          },
          "deposit_address": {
            "title": "Deposit Address",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "network": {
            "title": "Network",
            "type": "string"
          },
          "telegram_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telegram User Id"
          },
          "telegram_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telegram Username"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "display_name",
          "deposit_address",
          "network",
          "asset"
        ],
        "title": "MeOut",
        "type": "object"
      },
      "RegeneratedKeyOut": {
        "properties": {
          "api_key": {
            "title": "Api Key",
            "type": "string"
          }
        },
        "required": [
          "api_key"
        ],
        "title": "RegeneratedKeyOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "TG Pool administrator payment API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/telegram-users/{telegram_user_id}": {
      "get": {
        "operationId": "telegram_user_admin_telegram_users__telegram_user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "telegram_user_id",
            "required": true,
            "schema": {
              "title": "Telegram User Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-admin-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Telegram User"
      }
    },
    "/admin/telegram-users/{telegram_user_id}/balance": {
      "get": {
        "operationId": "telegram_user_balance_admin_telegram_users__telegram_user_id__balance_get",
        "parameters": [
          {
            "in": "path",
            "name": "telegram_user_id",
            "required": true,
            "schema": {
              "title": "Telegram User Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-admin-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Telegram User Balance"
      }
    },
    "/admin/telegram-users/{telegram_user_id}/regenerate-key": {
      "post": {
        "operationId": "regenerate_telegram_user_key_admin_telegram_users__telegram_user_id__regenerate_key_post",
        "parameters": [
          {
            "in": "path",
            "name": "telegram_user_id",
            "required": true,
            "schema": {
              "title": "Telegram User Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-admin-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegeneratedKeyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Regenerate Telegram User Key"
      }
    },
...
    "/admin/users": {
      "get": {
        "operationId": "admin_users_admin_users_get",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Search",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-admin-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUsersOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Users"
      },
      "post": {
        "operationId": "issue_user_admin_users_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-admin-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedUserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Issue User"
      }
    },
...
  }
}